Proof structs prefer vec of structs instead of paired vectors
Previously, MembershipProof and NonMembershipProof held two vectors of equal lengths, one holding NodeLabel and one holding Digest. This change introduces a Node struct that pairs these together, allowing the proofs to hold a single vector of these pairs.

The AppendOnlyProof was holding a single vector of tuples and has likewise been converted to prefer the Node struct.

In general, all NodeLabel+Digest pairings are now using the Node struct. The only one remaining is HistoryChildState::new which appears only to be used in tests. As such, I've let it be for now, just to keep this work internally contained to the proofs.