Change the types of AkdLabel and AkdValue to be an array of bytes.
https://github.com/novifinancial/akd/issues/104

Previously, these values were traded around as `String` values for convenience of readability. This change moves these values to `Vec<u8>` to reflect that these are intended as arbitrary blobs of bytes.

The POC continues to traffic in strings since it's a demonstration meant for the command line, where the readability is more convenient.