Make VRF(Public|Private) key Cloneable
This implements the std::clone::Clone trait on the VRFPublicKey and VRFPrivateKey structs. In the case of the PublicKey, the ed25519_dalek::PublicKey already supports clone so the trait can be #derived. The ed25519_dalek::SecretKey does not, so a manual implementation is required. It simply creates a new SecretKey out of the bytes of the original.