Add support for passing the epoch to the transaction operations
In some cases, we may want to offload the transaction to a backup storage as it's a snapshot of all of the changes occurring in that epoch. In this event, it's helpful to know **which** epoch a batch of changes is associated with. This is more future effort than actually applied in the repo, but provides downstream support for others.