Minimize over-the-wire encoding in protobuf formats.
This change removes trailing 0's from the node labels as they go over the wire, since we know the length will ALWAYS be 32 bytes. Therefore if there's a bunch of trailing 0's in the label, we can safely trim them off keeping only the most significant data (this is common with nodes higher up in the tree).

When decoding, we see how many bytes are there and put them in the leading spaces, in-order, with a simple slice copy.

Unit test coverage added in `akd_core/src/proto/tests.rs`