Lookup proof optional optimization
In the case of a lookup proof, generally we're normally doing up to `N = depth(Merkle Tree)` requests to preload the nodes for the proof. However this is still like 30+ queries on moderately sized trees (e.g. ~2B nodes)

In this PR, we introduce an optional optimization which will build all 256 labels which could be on the path from the root node to the intended leaf and try a single batch-get with all of them.

A lookup proof's `LookupInfo` needs 3 target labels to retrieve, so we're actually doing 3 * 256 node retrievals, but in a single query which is still often more optimal due to DB round-trip accesses than 32 targeted node requests.

After the first direct-path load for the 3 necessary targets, we then do a second batch-get of all of those node's children (which will only load the found node's children that aren't already in-cache) which gives us all of the sibling nodes to generate the necessary proof.