Add a Newtype Representative of a Read-only Directory (#359)
- Prior to this patch, we were reliant upon a boolean parameter passed to the
associated `new` fn on the `Directory` type to indicate whether the instance it
produced should be treated as a read-only entity. To more explicitly represent a
`Directory` which should be treated as a read-only entity, this patch introduces
a [newtype](https://doc.rust-lang.org/rust-by-example/generics/new_types.html) known
as `ReadOnlyDirectory` which acts as a simple wrapper for a `Directory`.
- To avoid confusion around whether or not a directory has the ability to publish,
the `ReadOnlyDirectory` and `Directory` types do not share any traits. Instead, the
`ReadOnlyDirectory` tpye wraps a `Directory` instance and essenitally calls into the
`Directory` for read-only operations. As such, we get compile time guarantees that a
`publish` operation is never possible via a `ReadOnlyDirectory` since the method simply
doesn't exist.

Co-authored-by: Dillon George <dgeo@fb.com>